🔨 refactor: my page 함수 및 컴포넌트명 변경
