+++ src/app/(user)/(service)/my-page/profile/page.tsx
@@ -7,7 +7,7 @@ import { ArrowRightIcon, SettingIcon } from "@/src/components/utils/icons";
 import { auth } from "@/src/auth";
 import { redirect } from "next/navigation";
 import Profile from "@/src/components/common/profile";
-import RecentDestinations from "@/src/features/my-page/my-page-direction";
+import RecentDestinations from "@/src/features/my-page/my-page-recent-path";
 
 export default async function Page() {
   const session = await auth();
